Player, Guard and system files
the guard script is the final AI script, which was built from scratch after giving up on the AI controller code. crouch, flashlight and pickup are all functions for the player. Jumpscare and jumpscare manager were intended to be used for an object to jumpscare the player, although this was never finished due to an error. Main menu, enablemouse, play and GameEnd are used to switch between scenes.